[CL-700] Move auth-owned layout components to CL ownership
