+++ apps/browser/src/popup/components/extension-anon-layout-wrapper/extension-anon-layout-wrapper.component.ts
@@ -1,76 +1,90 @@
 // FIXME: Update this file to be type safe and remove this and next line
 // @ts-strict-ignore
-import { ChangeDetectorRef, Component, OnDestroy, OnInit } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { Component, OnDestroy, OnInit } from "@angular/core";
 import { ActivatedRoute, Data, NavigationEnd, Router, RouterModule } from "@angular/router";
 import { Subject, filter, switchMap, takeUntil, tap } from "rxjs";
 
-import { AnonLayoutComponent } from "@bitwarden/auth/angular";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
-// This import has been flagged as unallowed for this class. It may be involved in a circular dependency loop.
-// eslint-disable-next-line no-restricted-imports
-import { Icon, Translation } from "@bitwarden/components";
+import {
+  Icon,
+  Icons,
+  IconModule,
+  Translation,
+  AnonLayoutComponent,
+  AnonLayoutWrapperData,
+  AnonLayoutWrapperDataService,
+} from "@bitwarden/components";
+import { I18nPipe } from "@bitwarden/ui-common";
 
-import { AnonLayoutWrapperDataService } from "./anon-layout-wrapper-data.service";
+import { CurrentAccountComponent } from "../../../auth/popup/account-switching/current-account.component";
+import { AccountSwitcherService } from "../../../auth/popup/account-switching/services/account-switcher.service";
+import { PopOutComponent } from "../../../platform/popup/components/pop-out.component";
+import { PopupHeaderComponent } from "../../../platform/popup/layout/popup-header.component";
+import { PopupPageComponent } from "../../../platform/popup/layout/popup-page.component";
 
-export interface AnonLayoutWrapperData {
-  /**
-   * The optional title of the page.
-   * If a string is provided, it will be presented as is (ex: Organization name)
-   * If a Translation object (supports placeholders) is provided, it will be translated
-   */
-  pageTitle?: string | Translation | null;
-  /**
-   * The optional subtitle of the page.
-   * If a string is provided, it will be presented as is (ex: user's email)
-   * If a Translation object (supports placeholders) is provided, it will be translated
-   */
-  pageSubtitle?: string | Translation | null;
-  /**
-   * The optional icon to display on the page.
-   */
-  pageIcon?: Icon | null;
-  /**
-   * Optional flag to either show the optional environment selector (false) or just a readonly hostname (true).
-   */
-  showReadonlyHostname?: boolean;
-  /**
-   * Optional flag to set the max-width of the page. Defaults to 'md' if not provided.
-   */
-  maxWidth?: "md" | "3xl";
-  /**
-   * Optional flag to set the max-width of the title area. Defaults to null if not provided.
-   */
-  titleAreaMaxWidth?: "md";
+export interface ExtensionAnonLayoutWrapperData extends AnonLayoutWrapperData {
+  showAcctSwitcher?: boolean;
+  showBackButton?: boolean;
+  showLogo?: boolean;
+  hideFooter?: boolean;
+  hideIcon?: boolean;
 }
 
 @Component({
-  templateUrl: "anon-layout-wrapper.component.html",
-  imports: [AnonLayoutComponent, RouterModule],
+  templateUrl: "extension-anon-layout-wrapper.component.html",
+  imports: [
+    AnonLayoutComponent,
+    CommonModule,
+    CurrentAccountComponent,
+    I18nPipe,
+    IconModule,
+    PopOutComponent,
+    PopupPageComponent,
+    PopupHeaderComponent,
+    RouterModule,
+  ],
 })
-export class AnonLayoutWrapperComponent implements OnInit, OnDestroy {
+export class ExtensionAnonLayoutWrapperComponent implements OnInit, OnDestroy {
   private destroy$ = new Subject<void>();
+
+  protected showAcctSwitcher: boolean;
+  protected showBackButton: boolean;
+  protected showLogo: boolean = true;
+  protected hideIcon: boolean = false;
 
   protected pageTitle: string;
   protected pageSubtitle: string;
   protected pageIcon: Icon;
   protected showReadonlyHostname: boolean;
   protected maxWidth: "md" | "3xl";
-  protected titleAreaMaxWidth: "md";
+  protected hasLoggedInAccount: boolean = false;
+  protected hideFooter: boolean;
+
+  protected theme: string;
+  protected logo = Icons.ExtensionBitwardenLogo;
 
   constructor(
     private router: Router,
     private route: ActivatedRoute,
     private i18nService: I18nService,
-    private anonLayoutWrapperDataService: AnonLayoutWrapperDataService,
-    private changeDetectorRef: ChangeDetectorRef,
+    private extensionAnonLayoutWrapperDataService: AnonLayoutWrapperDataService,
+    private accountSwitcherService: AccountSwitcherService,
   ) {}
 
-  ngOnInit(): void {
+  async ngOnInit(): Promise<void> {
     // Set the initial page data on load
     this.setAnonLayoutWrapperDataFromRouteData(this.route.snapshot.firstChild?.data);
+
     // Listen for page changes and update the page data appropriately
     this.listenForPageDataChanges();
     this.listenForServiceDataChanges();
+
+    this.accountSwitcherService.availableAccounts$
+      .pipe(takeUntil(this.destroy$))
+      .subscribe((accounts) => {
+        this.hasLoggedInAccount = accounts.some((account) => account.id !== "addAccount");
+      });
   }
 
   private listenForPageDataChanges() {
@@ -104,21 +118,37 @@ export class AnonLayoutWrapperComponent implements OnInit, OnDestroy {
       this.pageIcon = firstChildRouteData["pageIcon"];
     }
 
+    this.hideFooter = Boolean(firstChildRouteData["hideFooter"]);
     this.showReadonlyHostname = Boolean(firstChildRouteData["showReadonlyHostname"]);
     this.maxWidth = firstChildRouteData["maxWidth"];
-    this.titleAreaMaxWidth = firstChildRouteData["titleAreaMaxWidth"];
+
+    if (firstChildRouteData["showAcctSwitcher"] !== undefined) {
+      this.showAcctSwitcher = Boolean(firstChildRouteData["showAcctSwitcher"]);
+    }
+
+    if (firstChildRouteData["showBackButton"] !== undefined) {
+      this.showBackButton = Boolean(firstChildRouteData["showBackButton"]);
+    }
+
+    if (firstChildRouteData["showLogo"] !== undefined) {
+      this.showLogo = Boolean(firstChildRouteData["showLogo"]);
+    }
+
+    if (firstChildRouteData["hideIcon"] !== undefined) {
+      this.hideIcon = Boolean(firstChildRouteData["hideIcon"]);
+    }
   }
 
   private listenForServiceDataChanges() {
-    this.anonLayoutWrapperDataService
+    this.extensionAnonLayoutWrapperDataService
       .anonLayoutWrapperData$()
       .pipe(takeUntil(this.destroy$))
-      .subscribe((data: AnonLayoutWrapperData) => {
+      .subscribe((data: ExtensionAnonLayoutWrapperData) => {
         this.setAnonLayoutWrapperData(data);
       });
   }
 
-  private setAnonLayoutWrapperData(data: AnonLayoutWrapperData) {
+  private setAnonLayoutWrapperData(data: ExtensionAnonLayoutWrapperData) {
     if (!data) {
       return;
     }
@@ -139,13 +169,29 @@ export class AnonLayoutWrapperComponent implements OnInit, OnDestroy {
       this.pageIcon = data.pageIcon !== null ? data.pageIcon : null;
     }
 
+    if (data.hideFooter !== undefined) {
+      this.hideFooter = data.hideFooter !== null ? data.hideFooter : null;
+    }
+
     if (data.showReadonlyHostname !== undefined) {
       this.showReadonlyHostname = data.showReadonlyHostname;
     }
 
-    // Manually fire change detection to avoid ExpressionChangedAfterItHasBeenCheckedError
-    // when setting the page data from a service
-    this.changeDetectorRef.detectChanges();
+    if (data.showAcctSwitcher !== undefined) {
+      this.showAcctSwitcher = data.showAcctSwitcher;
+    }
+
+    if (data.showBackButton !== undefined) {
+      this.showBackButton = data.showBackButton;
+    }
+
+    if (data.showLogo !== undefined) {
+      this.showLogo = data.showLogo;
+    }
+
+    if (data.hideIcon !== undefined) {
+      this.hideIcon = data.hideIcon;
+    }
   }
 
   private handleStringOrTranslation(value: string | Translation): string {
@@ -163,8 +209,11 @@ export class AnonLayoutWrapperComponent implements OnInit, OnDestroy {
     this.pageSubtitle = null;
     this.pageIcon = null;
     this.showReadonlyHostname = null;
+    this.showAcctSwitcher = null;
+    this.showBackButton = null;
+    this.showLogo = null;
     this.maxWidth = null;
-    this.titleAreaMaxWidth = null;
+    this.hideFooter = null;
   }
 
   ngOnDestroy() {
