+++ libs/auth/src/angular/icons/index.ts
@@ -1,8 +1,4 @@
-export * from "./bitwarden-logo.icon";
-export * from "./bitwarden-shield.icon";
 export * from "./devices.icon";
-export * from "./lock.icon";
-export * from "./registration-check-email.icon";
 export * from "./user-lock.icon";
 export * from "./user-verification-biometrics-fingerprint.icon";
 export * from "./wave.icon";
