+++ libs/components/src/anon-layout/anon-layout.mdx
@@ -6,8 +6,8 @@ import * as stories from "./anon-layout.stories";
 
 # AnonLayout Component
 
-The Auth-owned AnonLayoutComponent is to be used primarily for unauthenticated pages\*, where we
-don't know who the user is.
+The AnonLayoutComponent is to be used primarily for unauthenticated pages\*, where we don't know who
+the user is.
 
 \*There will be a few exceptions to this&mdash;that is, AnonLayout will also be used for the Unlock
 and View Send pages.
