+++ libs/components/src/anon-layout/index.ts
@@ -1,0 +1,4 @@
+export * from "./anon-layout-wrapper-data.service";
+export * from "./anon-layout-wrapper.component";
+export * from "./anon-layout.component";
+export * from "./default-anon-layout-wrapper-data.service";
