+++ libs/components/src/icon/icons/index.ts
@@ -1,8 +1,13 @@
-export * from "./search";
-export * from "./security";
+export * from "./bitwarden-logo.icon";
+export * from "./bitwarden-shield.icon";
+export * from "./extension-bitwarden-logo.icon";
+export * from "./lock.icon";
+export * from "./generator";
 export * from "./no-access";
 export * from "./no-results";
-export * from "./generator";
+export * from "./registration-check-email.icon";
+export * from "./search";
+export * from "./security";
 export * from "./send";
 export * from "./settings";
 export * from "./vault";
