+++ libs/components/src/index.ts
@@ -1,5 +1,6 @@
 export { ButtonType, ButtonLikeAbstraction } from "./shared/button-like.abstraction";
 export * from "./a11y";
+export * from "./anon-layout";
 export * from "./async-actions";
 export * from "./avatar";
 export * from "./badge-list";
